Completely remove Java agent from sample project
This is a special commit for Tibor Digana in order to make him believe
that Java agents have nothing whatsoever to do with this problem, but
simply logging to the console in the forked JVM running the test does.

The agent stuff is still on branch 'use-agent', so we can continue to
test and debug there after the basic problem with using
SurefireForkNodeFactory has been solved (see SUREFIRE-1881).